+++ pkg/execution/mutation/patcher_jobconfigs.go
@@ -21,7 +21,7 @@ import (
 
 	execution "github.com/furiko-io/furiko/apis/execution/v1alpha1"
 	"github.com/furiko-io/furiko/pkg/runtime/controllercontext"
-	"github.com/furiko-io/furiko/pkg/utils/webhook"
+	"github.com/furiko-io/furiko/pkg/runtime/webhook"
 )
 
 // JobConfigPatcher encapsulates high-level patch methods for JobConfigs.
