+++ pkg/execution/webhooks/jobmutatingwebhook/webhook.go
@@ -38,8 +38,8 @@ import (
 	"github.com/furiko-io/furiko/pkg/runtime/controllercontext"
 	"github.com/furiko-io/furiko/pkg/runtime/controllermanager"
 	"github.com/furiko-io/furiko/pkg/runtime/controllerutil"
+	"github.com/furiko-io/furiko/pkg/runtime/webhook"
 	"github.com/furiko-io/furiko/pkg/utils/cmp"
-	"github.com/furiko-io/furiko/pkg/utils/webhook"
 )
 
 const (
